% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthoPanels-package.R
\name{BES_panel}
\alias{BES_panel}
\title{Responses from the 2010 British Election Study}
\format{A data frame with 5535 rows and 11 variables}
\description{
A survey of 1845 respondents using 3 waves of panel survey data
from the 2010 British Election Study. The variables are as follows:
}
\details{
\itemize{
  \item n case number
  \item t time wave
  \item Econ Assessment of change in the national economic situation over the past 12 months
        (1-5, 1=\sQuote{got a lot worse}, 5=\sQuote{got a lot better})
  \item Clegg Evaluation of Liberal Party leader Nick Clegg
        (0-10, 0=\sQuote{strongly dislike} and 10=\sQuote{strongly like})
  \item Brown Evaluation of Labour Party leader Gordon Brown
  \item Cameron Evaluation of Conservative Party leader David Cameron
  \item Approve Approval of the government, as expressed by feeling about the ruling Labour Party
        (0-10, 0=\sQuote{strongly dislike}, 10=\sQuote{strongly like})
  \item NHS Assesment of the current government's handling of the National Health
        Service (1-5, 1=\sQuote{very badly}, 5=\sQuote{very well})
  \item Terror Assesment of the current government's handling of terrorism
        (1-5, 1=\sQuote{very badly}, 5=\sQuote{very well})
  \item PID Personal identification with the Labour Party
        (0/1, 0=\sQuote{no}, 1=\sQuote{yes})
  \item Tax Preference for policy on taxes and health and social spending
        (0-10, 0=\sQuote{cut taxes a lot and spend much less},
        10=\sQuote{increase taxes a lot and spend much more})
}
}

