% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{p_rho}
\alias{p_rho}
\title{Returns the posterior density \eqn{p(\rho|\Theta)}}
\usage{
p_rho(x, y, rho, log.p = FALSE)
}
\arguments{
\item{x}{an array of dimension \code{time x variable x case} of terms.}

\item{y}{a matrix of dimension \code{time x case} of responses.}

\item{rho}{vector of quantiles.}

\item{log.p}{if \code{TRUE}, probabilities are given as \code{log(p)}.}
}
\description{
Returns the posterior density \eqn{p(\rho|\Theta)}
}

